Module stream_httparse::header

Holds some more Types that are needed for Headers

Structs

Header
A single HTTP-Header Pair(Key-Value)

Enums

HeaderKey
Allows the HeaderKey to take the form of a variety of different valid Types, mostly related to their lifetimes. This however also gives more control over how they are compared to each other, ignoring case in this case

HeaderValue
A single HeaderValue that can hold Data in a variety of forms allowing for easier and more flexible use