Module stream_httparse::streaming_parser

This module holds all the Parsers that can deal with parsing the Data in multiple chunks and dont need all of it right away

Structs

ChunkParser

A single ChunkParser instance used to parse multiple Chunks one after the other

ReqParser

A single Instance of the RequestParser that is used to parse HTTP-Requests

RespParser

A single Instance of the ResponseParser that is used to parse HTTP-Responses

Enums

ParseError

Indicates all Errors related to parsing a Request/Response/Chunk

Type Definitions

ParseResult

A simple Result wrapper that already has the ParseError type as its Error-Type