Struct stream_httparse::Response

pub struct Response<'a> { /* fields omitted */ }

Represents a single HTTP-Request

Implementations

impl<'a> Response<'a>

pub fn new(
protocol: &'a str,
status_code: StatusCode,
headers: Headers<'a>,
body: Vec<u8>
) -> Self

Creates a new Response with the given Data as its inital State

pub fn serialize(&self) -> (Vec<u8>, &[u8])

Serialzes the Response and returns the Data as a tuple of form (HTTP-Head, HTTP-Body)

pub fn protocol(&self) -> &str

Returns the Protocol of the Response

pub fn status_code(&self) -> &StatusCode

Returns the StatusCode of the Response

pub fn headers(&self) -> &Headers<'a>

Returns the Headers of the Response

pub fn body(&self) -> &[u8]

Returns the Body of the Response

pub fn add_header<'b, K, V>(&mut self, key: K, value: V) where
'b: 'a,
K: Into<HeaderKey<'a>>,
V: Into<HeaderValue<'a>>,

Adds the Key-Value Pair as a new Header to the Response or replaces the old Value of the Header if it already existed on the Response

pub fn set_body(&mut self, n_body: Vec<u8>)

Replaces the old Body of the Response with the new given Body and updates the Content-Length Header as well with the new Length

pub fn is_chunked(&self) -> bool

Checks if the Response is send using Transfer-Encoding: Chunked

pub fn to_owned<'refed, 'owned>(&'refed self) -> Response<'owned>

Clones the entire Response to produce a new indepandent Response